requirements.txt
Jinja2==2.10
lxml==4.2.5
numpy==1.14.3
pandas==0.23.0
requests==2.18.1
selenium==3.8.0
urllib3==1.21.1
webkit2png==0.8.2
Pillow==5.1.0
ipython==6.5.0
webdriver-manager==1.7
html5lib==1.0.1
bs4==0.0.1
mypy==0.630
